#5.2 Update Profile Bugfix

+++ src/components/Dweet.js
@@ -9,10 +9,15 @@ const Dweet = ({ dweetObj, isOwner }) => {
     const ok = window.confirm("Are you sure you want to delete this dweet?");
 
     if (ok) {
-      await dbService.doc(`dweets/${dweetObj.id}`).delete();
-      //   await dbService.collection("dweets").doc(`${dweetObj.id}`).delete();
-
-      await storageService.refFromURL(dweetObj.attachmentUrl).delete();
+      try {
+        await dbService.doc(`dweets/${dweetObj.id}`).delete();
+        //   await dbService.collection("dweets").doc(`${dweetObj.id}`).delete();
+        if (dweetObj.attachmentUrl) {
+          await storageService.refFromURL(dweetObj.attachmentUrl).delete();
+        }
+      } catch (error) {
+        console.log(error);
+      }
     }
   };
